Refactor and document list processing methods
Added comments to explain the use of `All`, `Any`, and `Chunk` methods. Implemented the `All` method to check if all numbers are greater than 2 and dumped the result. Commented out the `Any` method for numbers greater than 15. Removed and commented out the `Chunk` method for dividing the list into parts of 3 elements.